refactor(fs): pluralise backend folder names
- Renamed backend/room to backend/rooms
- Renamed backend/space to backend/spaces
- Updated import paths in API routes to use new folder names
- Fixed critical imports for compilation

Note: Additional import path fixes may be needed in backend files
