+++ app/api/room/reservation/(adaptor)/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { SbRsvRepository } from '@/backend/common/infrastructure/repositories/SbRsvRepository';
-import { GetRoomRsvUsecase } from '@/backend/room/reservation/usecases/GetRoomRsvUsecase';
+import { GetRoomRsvUsecase } from '@/backend/rooms/reservation/usecases/GetRoomRsvUsecase';
 import { createClient } from '@/backend/common/infrastructure/supabase/server';
 
 export async function GET(request: NextRequest) {
